web/requirements.txt
asgiref==3.3.2
baycomp==1.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
Django==3.2
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-gdpr-assist
idna==2.10
install==1.3.4
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
oauthlib==3.1.0
pandarallel==1.5.2
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.0b2
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python3-openid==3.2.0
pytz==2021.1
ranky==0.2.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scikit-survival==0.15.0.post0
scipy
seaborn==0.11.1
six==1.15.0
sqlparse==0.4.1
statsmodels
threadpoolctl==2.1.0
tqdm==4.60.0
urllib3==1.26.4
gunicorn==19.3.0
psycopg2-binary==2.8.5
arrow==0.7.0
click==6.2
django-rq==2.4.1
redis==3.5.0
rq==1.8.1
plotly==4.14.3
lifelines
